[SW-2146] Cleanup in tests -> move to right ai.h2o.sparkling packages
