+++ build.gradle
@@ -47,7 +47,11 @@ ext {
     project(':sparkling-water-extensions')
   ]
   // Projects with integration tests
-  integTestProjects = [project(':sparkling-water-core'), project(':sparkling-water-examples')]
+  integTestProjects = [
+    project(':sparkling-water-core'),
+    project(':sparkling-water-examples'),
+    project(':sparkling-water-ml')
+  ]
 
   // Projects with benchmarks
   benchProjects = [project(':sparkling-water-core'), project(':sparkling-water-benchmarks')]
