+++ core/src/bench/scala/ai/h2o/sparkling/bench/BenchResult.scala
@@ -14,19 +14,23 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package water.api
 
-import java.io.File
+package ai.h2o.sparkling.bench
 
-object TestUtils {
+import scala.concurrent.duration.TimeUnit
 
-  def locate(name: String): String = {
-    val file = new File("./examples/" + name)
-    if (file.exists()) {
-      file.getAbsolutePath
-    } else {
-      // testing from IDEA
-      new File("../examples/" + name).getAbsolutePath
-    }
+case class BenchResult(mean: Float, stdDev: Float, min: Float, max: Float, unit: TimeUnit) {
+  def show(): String = {
+    f"${mean}%4f ± ${stdDev}%4f (${min}%4f, ${max}%4f)"
+  }
+}
+
+object BenchResult {
+  def apply(measurements: Array[Long], inputUnit: TimeUnit, outputUnit: TimeUnit): BenchResult = {
+    val convMeasurements = measurements.map(x => outputUnit.convert(x, inputUnit))
+    val mean = convMeasurements.sum.toFloat / convMeasurements.length
+    val stdev =
+      (Math.sqrt(convMeasurements.map(x => (x - mean) * (x - mean)).sum / (convMeasurements.length - 1))).toFloat
+    new BenchResult(mean, stdev, convMeasurements.min, convMeasurements.max, outputUnit)
   }
 }
