+++ core/src/bench/scala/ai/h2o/sparkling/bench/BenchSuite.scala
@@ -19,8 +19,9 @@ package ai.h2o.sparkling.bench
 
 import java.util.concurrent.TimeUnit
 
-import org.apache.spark.h2o.utils.BenchUtils.bench
 import org.scalatest.FunSuite
+
+import scala.concurrent.duration.TimeUnit
 
 class BenchSuite extends FunSuite {
 
@@ -37,5 +38,39 @@ class BenchSuite extends FunSuite {
     }
 
     registerTest(testName)(body)
+  }
+
+  /**
+    * Measure execution time of given block in nanoseconds.
+    *
+    * @param block block to measure
+    * @return number of ns to execute given block
+    */
+  def timer(block: => Unit): Long = {
+    val now = System.nanoTime()
+    block
+    System.nanoTime() - now
+  }
+
+  /**
+    * Benchmark given block of code.
+    *
+    * @param iterations number of iterations to execute the block of code
+    * @param block      block to execute as benchmark
+    * @return
+    */
+  def bench(iterations: Int, warmup: Int = 4, outputTimeUnit: TimeUnit = TimeUnit.MILLISECONDS)(
+      block: => Unit): BenchResult = {
+    val times = new Array[Long](iterations)
+    // Warmup
+    for (i <- 0 until warmup) {
+      timer(block)
+    }
+    // Measure
+    for (i <- 0 until iterations) {
+      times(i) = timer(block)
+    }
+
+    BenchResult(times, TimeUnit.NANOSECONDS, outputTimeUnit)
   }
 }
