+++ core/src/integTest/scala/ai/h2o/sparkling/internal/H2OContextConversionOnSubsetExecutors.scala
@@ -16,10 +16,10 @@
  */
 package ai.h2o.sparkling.internal
 
+import ai.h2o.sparkling.TestUtils.DoubleHolder
 import ai.h2o.sparkling.backend.internal.InternalBackendConf
+import ai.h2o.sparkling.{SharedH2OTestContext, TestUtils}
 import org.apache.spark.SparkContext
-import org.apache.spark.h2o.utils.{SharedH2OTestContext, TestFrameUtils}
-import org.apache.spark.h2o.testdata.DoubleHolder
 import org.junit.runner.RunWith
 import org.scalatest.FunSuite
 import org.scalatest.junit.JUnitRunner
@@ -42,7 +42,7 @@ class H2OContextConversionOnSubsetExecutors extends FunSuite with SharedH2OTestC
     val rdd = sc.parallelize(1 to 1000, 100).map(v => Some(v))
     val h2oFrame = hc.asH2OFrame(rdd)
 
-    TestFrameUtils.assertBasicInvariants(rdd, h2oFrame, (rowIdx, vec) => {
+    TestUtils.assertBasicInvariants(rdd, h2oFrame, (rowIdx, vec) => {
       val nextRowIdx = rowIdx + 1
       val value = vec.at(rowIdx)
       assert(nextRowIdx == value, "The H2OFrame values should match row numbers+1")
@@ -58,7 +58,7 @@ class H2OContextConversionOnSubsetExecutors extends FunSuite with SharedH2OTestC
 
     val convertedDf = hc.asDataFrame(hf)
     import spark.implicits._
-    TestFrameUtils.assertDataFramesAreIdentical(originalRdd.toDF, convertedDf.toDF())
+    TestUtils.assertDataFramesAreIdentical(originalRdd.toDF, convertedDf.toDF())
   }
 
   test("asRDD conversion on subset of executors") {
@@ -69,6 +69,6 @@ class H2OContextConversionOnSubsetExecutors extends FunSuite with SharedH2OTestC
 
     val convertedRdd = hc.asRDD[DoubleHolder](hf)
 
-    TestFrameUtils.assertDataFramesAreIdentical(originalRdd.toDF, convertedRdd.toDF())
+    TestUtils.assertDataFramesAreIdentical(originalRdd.toDF, convertedRdd.toDF())
   }
 }
